+++ NEWS.md
@@ -27,6 +27,7 @@ Version 0.16.4b1 - dev
 - BUGFIX: correct application of the `Dimension.dxf.insert` attribute
 - BUGFIX: fixed incorrect "thickness" transformation of OCS entities
 - BUGFIX: add missing "width" transformation to POLYLINE and LWPOLYLINE
+- BUGFIX: drawing add-on handles the invisible flag for INSERT correct
 
 Version 0.16.3 - 2021-05-22
 ---------------------------